+++ lib/matrix_homeserver_recommendations.dart
@@ -7,3 +7,5 @@ export 'src/models/homeserver_benchmark_result.dart';
 
 export 'src/parsers/joinmatrix_org/joinmatrix_org_parser.dart';
 export 'src/parsers/joinmatrix_org/joinmatrix_org_server.dart';
+
+export 'src/parsers/local_homeserver_parser/local_homeserver_parser.dart';