+++ test/matrix_homeserver_recommendations_test.dart
@@ -17,4 +17,31 @@ void main() {
       expect(benchmark.isNotEmpty, isTrue);
     });
   });
+
+  group('local homeserver parser', () {
+    test('NO WAR!!!', () async {
+      var server = LocalHomeserverParser('ua').fetchHomeservers().first;
+      expect(server.baseUrl.host, 'gemeinsam.jetzt');
+    });
+
+    test('United States', () async {
+      var server = LocalHomeserverParser('us').fetchHomeservers().first;
+      expect(server.baseUrl.host, 'buyvm.net');
+    });
+
+    test('Russia', () async {
+      var server = LocalHomeserverParser('ru').fetchHomeservers().first;
+      expect(server.baseUrl.host, 'rumatrix.org');
+    });
+
+    test('Germany', () async {
+      var server = LocalHomeserverParser('de').fetchHomeservers().first;
+      expect(server.baseUrl.host, 'envs.de');
+    });
+
+    test('Invalid code', () async {
+      var server = LocalHomeserverParser('tlh').fetchHomeservers();
+      expect(server.isEmpty, isTrue);
+    });
+  });
 }